+++ pya/aserver.py
@@ -1,8 +1,11 @@
+from .helper.backend import determine_backend
 import copy
-import time
 import logging
-import numpy as np
+import time
+from typing import Optional, Union
 from warnings import warn
+
+import numpy as np
 
 
 _LOGGER = logging.getLogger(__name__)
@@ -48,8 +51,9 @@ class Aserver:
         else:
             warn("Aserver:shutdown_default_server: no default_server to shutdown")
 
-    def __init__(self, sr=44100, bs=None, device=None,
-                 channels=None, backend=None, **kwargs):
+    def __init__(self, sr: int = 44100, bs: Optional[int] = None,
+                 device: Optional[int] = None, channels: Optional[int] = None,
+                 backend=None, **kwargs):
         """Aserver manages an pyaudio stream, using its aserver callback
         to feed dispatched signals to output at the right time.
 
@@ -63,7 +67,7 @@ class Aserver:
             The device index based on pya.device_info(), default is None which will set 
             the default device from PyAudio
         channels : int
-            number of channel (Default value = 2)
+            number of channel, default is the max output channels of the device
         kwargs : backend parameter
 
         Returns
@@ -73,11 +77,8 @@ class Aserver:
         """
         # TODO check if channels is overwritten by the device.
         self.sr = sr
-        if backend is None:
-            from .backend.PyAudio import PyAudioBackend
-            self.backend = PyAudioBackend(**kwargs)
-        else:
-            self.backend = backend
+        self.stream = None
+        self.backend = determine_backend(**kwargs) if backend is None else backend
         self.bs = bs or self.backend.bs
         # Get audio devices to input_device and output_device
         self.input_devices = []
@@ -88,55 +89,54 @@ class Aserver:
             if int(self.backend.get_device_info_by_index(i)['maxOutputChannels']) > 0:
                 self.output_devices.append(self.backend.get_device_info_by_index(i))
 
-        self._device = device or self.backend.get_default_output_device_info()['index']
-        self.channels = channels or self.backend.get_device_info_by_index(self.device)['maxOutputChannels']
+        self._device = self.backend.get_default_output_device_info()['index'] if device is None else device
+        self._channels = channels or self.max_out_chn
 
         self.gain = 1.0
         self.srv_onsets = []
-        self.srv_asigs = []
         self.srv_curpos = []  # start of next frame to deliver
+        self.srv_asigs = []
         self.srv_outs = []  # output channel offset for that asig
-        self.stream = None
         self.boot_time = 0  # time.time() when stream starts
         self.block_cnt = 0  # nr. of callback invocations
         self.block_duration = self.bs / self.sr  # nominal time increment per callback
         self.block_time = 0  # estimated time stamp for current block
         self._stop = True
-        self.empty_buffer = np.zeros((self.bs, self.channels),
-                                     dtype=self.backend.dtype)
+        self.empty_buffer = np.zeros((self.bs, self.channels), dtype=self.backend.dtype)
         self._is_active = False
 
     @property
+    def channels(self):
+        return self._channels
+
+    @channels.setter
+    def channels(self, val: int):
+        """
+        Set the number of channels. Aserver needs reboot.
+        """
+        if val > self.max_out_chn:
+            raise ValueError(f"AServer: channels {val} > max {self.max_out_chn}")
+        self._channels = val
+
+    @property
     def device_dict(self):
         return self.backend.get_device_info_by_index(self._device)
 
     @property
-    def max_out_chn(self):
+    def max_out_chn(self) -> int:
         return int(self.device_dict['maxOutputChannels'])
 
     @property
-    def max_in_chn(self):
+    def max_in_chn(self) -> int:
         return int(self.device_dict['maxInputChannels'])
-
-    @property
-    def device_dict(self):
-        return self.backend.get_device_info_by_index(self._device)
-
-    @property
-    def max_out_chn(self):
-        return int(self.device_dict['maxOutputChannels'])
-
-    @property
-    def max_in_chn(self):
-        return int(self.device_dict['maxInputChannels'])
-
-    @property
-    def device(self):
-        return self._device
 
     @property
     def is_active(self) -> bool:
         return self.stream is not None and self.stream.is_active()
+
+    @property
+    def device(self):
+        return self._device
 
     @device.setter
     def device(self, val):
@@ -146,12 +146,11 @@ class Aserver:
             self.channels = self.max_out_chn
 
     def __repr__(self):
-        state = False
         msg = f"""AServer: sr: {self.sr}, blocksize: {self.bs},
          Stream Active: {self.is_active}, Device: {self.device_dict['name']}, Index: {self.device_dict['index']}"""
         return msg
 
-    def get_devices(self, verbose=False):
+    def get_devices(self, verbose: bool = False):
         """Return (and optionally print) available input and output device"""
         if verbose:
             print("Input Devices: ")
@@ -162,8 +161,9 @@ class Aserver:
              for i in self.output_devices]
         return self.input_devices, self.output_devices
 
-    def set_device(self, idx, reboot=True):
-        """Set audio device
+    def set_device(self, idx: int, reboot: bool = True):
+        """Set audio device, an alternative way is to direct set the device property, i.e. Aserver.device = 1, 
+        but that will not reboot the server.
 
         Parameters
         ----------
@@ -215,8 +215,16 @@ class Aserver:
         self.stream = None
         return 0
 
-    def play(self, asig, onset=0, out=0, **kwargs):
-        """Dispatch asigs or arrays for given onset."""
+    def play(self, asig, onset: Union[int, float] = 0, out: int = 0, **kwargs):
+        """Dispatch asigs or arrays for given onset.
+
+        asig: pya.Asig
+            An Asig object
+        onset: int or float
+            Time when the sound should play, 0 means asap
+        out: int
+            Output channel
+        """
         self._stop = False
 
         sigid = id(asig)  # for copy check
@@ -316,5 +324,3 @@ class Aserver:
         self.quit()
         self.backend.terminate()
 
-    def __enter__(self):
-        return self.boot()